// oa-hedera/oa-verify/src/types/core.ts
export interface VerifierOptions {
  client: Client;
  resolver?: Resolver;
}
// ...
export interface VerificationFragment {
  name: string;
  type: VerificationFragmentType;
  status: VerificationFragmentStatus;
}
// ...
export interface SkippedVerificationFragment extends VerificationFragment {
  status: "SKIPPED";
  reason: Reason;
  // this is useless, it's just to make it easier to work with fragment, otherwise data is not available and a consumer needs to narrow the type down first
  data?: never;
}
export type VerificationFragmentType = "DOCUMENT_INTEGRITY" | "DOCUMENT_STATUS" | "ISSUER_IDENTITY";
export type VerificationFragmentStatus = "ERROR" | "VALID" | "INVALID" | "SKIPPED";
// ...
export interface Verifier<V extends VerificationFragment> {
  skip: (document: DocumentsToVerify, options: VerifierOptions) => Promise<SkippedVerificationFragment>;
  test: (document: DocumentsToVerify, options: VerifierOptions) => boolean;
  verify: (document: DocumentsToVerify, options: VerifierOptions) => Promise<V>;
}
export type Hash = string;

export type DocumentsToVerify =
  | WrappedDocument<v2.OpenAttestationDocument>
  | WrappedDocument<v3.OpenAttestationDocument>
  | SignedWrappedDocument<v2.OpenAttestationDocument>;